+++ src/provider/aws/iam/groups/cmds/hasManagedAdmin.ts
@@ -1,24 +1,25 @@
+/** TODO: this needs to support resource */
+
 import { AuditResultInterface, AWSScannerInterface } from 'cloud-search'
 import { AWS } from '../../../../../lib/aws/AWS'
-
 const rule = 'HasManagedAdmin'
 
 export const command = `${rule} [args]`
 
-export const desc = `Users should not have the AWS managed administrator policy 
-directly applied to the user's permissions. It is important to understand 
-that this rule only checks users to see if the AdministratorAccess managed 
-policy is directly attached.  A user could still have admin via an inline or 
-group policy. 
+export const desc = `User groups should not have the AWS managed administrator 
+policy directly attached. It is important to understand that only checks user
+groups. A user could still have admin via an inline or a directly attached
+policy. 
 
-  OK      - User does not have AdministratorAccess policy directly attached 
-  UNKNOWN - Unable to determine if the user has AdministratorAccess directly attached 
-  FAIL    - User has AdministratorAccess policy directly attached
+  OK      - Group does not have AdministratorAccess policy attached 
+  UNKNOWN - Unable to determine if the group has AdministratorAccess attached 
+  FAIL    - Group has AdministratorAccess policy attached
 
-  note: this rule does not check inline policy.  It is still possible that a user has administrator rights if they
-        were applied via a inline policy. 
+  note: This rule does not check inline policy.  It is still possible that a 
+        group has administrator rights if they were applied via a inline policy. 
 
-  note: this rule does not check group policy.  It is still possible the user is in a group that has admin attached
+  note: This rule does not check direct user policy.  It is still possible the 
+        user has admin directly attached.
 
 `
 
@@ -45,7 +46,7 @@ export default class HasManagedAdmin extends AWS {
       time: now.toISOString(),
     }
 
-    const policies = await this.listAttachedUserPolicies(resource)
+    const policies = await this.listAttachedGroupPolicies(resource)
 
     for (const policy of policies) {
       if (policy.PolicyName === 'AdministratorAccess') {
@@ -64,10 +65,10 @@ export default class HasManagedAdmin extends AWS {
     if (resourceId) {
       await this.audit({ resource: resourceId })
     } else {
-      const users = await this.listUsers()
+      const groups = await this.listGroups()
 
-      for (const user of users) {
-        await this.audit({ resource: user.UserName })
+      for (const group of groups) {
+        await this.audit({ resource: group.GroupName })
       }
     }
   }