+++ src/provider/aws/iam/roles/cmds/hasManagedAdmin.ts
@@ -5,20 +5,17 @@ const rule = 'HasManagedAdmin'
 
 export const command = `${rule} [args]`
 
-export const desc = `Users should not have the AWS managed administrator policy 
-directly applied to the user's permissions. It is important to understand 
-that this rule only checks users to see if the AdministratorAccess managed 
-policy is directly attached.  A user could still have admin via an inline or 
-group policy. 
+export const desc = `Roles should not have the AWS managed administrator policy 
+directly applied. It is important to understand that this rule only checks roles
+to see if the AdministratorAccess managed policy is directly attached.  A role
+could still have admin via an inline policy.
 
-  OK      - User does not have AdministratorAccess policy directly attached 
-  UNKNOWN - Unable to determine if the user has AdministratorAccess directly attached 
-  FAIL    - User has AdministratorAccess policy directly attached
+  OK      - Role does not have AdministratorAccess policy directly attached 
+  UNKNOWN - Unable to determine if the role has AdministratorAccess directly attached 
+  FAIL    - Role has AdministratorAccess policy directly attached
 
-  note: this rule does not check inline policy.  It is still possible that a user has administrator rights if they
-        were applied via a inline policy. 
-
-  note: this rule does not check group policy.  It is still possible the user is in a group that has admin attached
+  note: this rule does not check inline policy.  It is still possible that a 
+        role has administrator rights if they were applied via a inline policy. 
 
 `
 
@@ -45,7 +42,7 @@ export default class HasManagedAdmin extends AWS {
       time: now.toISOString(),
     }
 
-    const policies = await this.listAttachedUserPolicies(resource)
+    const policies = await this.listAttachedRolePolicies(resource)
 
     for (const policy of policies) {
       if (policy.PolicyName === 'AdministratorAccess') {
@@ -64,10 +61,10 @@ export default class HasManagedAdmin extends AWS {
     if (resourceId) {
       await this.audit({ resource: resourceId })
     } else {
-      const users = await this.listUsers()
+      const roles = await this.listRoles()
 
-      for (const user of users) {
-        await this.audit({ resource: user.UserName })
+      for (const role of roles) {
+        await this.audit({ resource: role.RoleName })
       }
     }
   }