+++ CHANGELOG.md
@@ -7,7 +7,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.6.0] - 2020-10-26
+
+### Added
+
 - AWS - iam, groups - HasManagedAdmin
+- AWS - iam, roles - HasManagedAdmin
+- AWS - guardduty - DetectorEnabled
+- AWS - guardduty - DetectorExists
+- AWS - guardduty - DetectorDataSources
+
+### Changed
+
+- fixed region reference as `all` when it should have been `global`
+
+### Removed
 
 ## [1.5.0] - 2020-10-24